[phoenix05] 1.增加phoenix05芯片处理代码
