+++ tcl/board/phoenix05.cfg
@@ -6,5 +6,5 @@ source [find interface/ftdi/openocd-usb.cfg]
 transport select jtag
 adapter speed 4000
 
-set CHIPNAME phoenix
+set CHIPNAME phoenix05
 source [find target/phoenix05.cfg]
